Tomaszzukowski develop feature UI for raports

+++ Solution1/ProjectClock.MVC/Controllers/AccountController.cs
@@ -39,6 +39,7 @@ namespace ProjectClock.MVC.Controllers
 
             if (!resultDto.LoginWasSuccessful)
             {
+                dto.LoginWasSuccessful = false;
                 return View(dto);
             }
 
@@ -60,17 +61,17 @@ namespace ProjectClock.MVC.Controllers
         [AllowAnonymous]
         public async Task<IActionResult> Register(RegisterDto dto)
         {
-            //if (!ModelState.IsValid)
-            //{
-            //    return View(dto);
-            //}
+            if (!ModelState.IsValid)
+            {
+                return View(dto);
+            }
 
-            
+
             var resultDto = await _accountService.RegisterAccount(dto);
-            
+            dto.Results = resultDto;
 
             if (resultDto.RegistrationFailed)
-            {
+            {             
                 return View(dto);
             }
 
@@ -82,26 +83,13 @@ namespace ProjectClock.MVC.Controllers
         {
             await HttpContext.SignOutAsync(CookieAuthenticationDefaults.AuthenticationScheme);
 
-            return RedirectToAction("account", "Login");
+            return RedirectToAction("Login", "Account");
         }
 
         [Authorize(Roles = "User")]
         public async Task<IActionResult> EditEmail()
         {
-            if (!HttpContext.User.Claims.TryGetAuthenticatedUserId(out var userId))
-            {
-                return RedirectToAction("Index", "Home");
-            }
-
-            var accountEmail = await _accountService.GetAccountEmail(userId);
-
-            var dto = new EditEmailDto
-            {
-                Id = userId,
-                CurrentEmail = accountEmail
-            };
-
-            return View(dto);
+            return View();
         }
 
         [HttpPost]
@@ -112,12 +100,19 @@ namespace ProjectClock.MVC.Controllers
             {
                 return View(dto);
             }
+            if (!HttpContext.User.Claims.TryGetAuthenticatedUserId(out var userId))
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
+            dto.Id = userId;
 
             var resultDto = await _accountService.EditAccountEmail(dto);
 
             if (resultDto.EditEmailFailed)
             {
-                return View(resultDto);
+                dto.Result = resultDto;
+                return View(dto);
             }
 
             return RedirectToAction("Index", "Home");
@@ -126,17 +121,7 @@ namespace ProjectClock.MVC.Controllers
         [Authorize(Roles = "User")]
         public ActionResult EditPassword()
         {
-            if (!HttpContext.User.Claims.TryGetAuthenticatedUserId(out var userId))
-            {
-                return RedirectToAction("Index", "Home");
-            }
-
-            var dto = new EditPasswordDto
-            {
-                UserId = userId
-            };
-
-            return View(dto);
+            return View();
         }
 
         [HttpPost]
@@ -147,12 +132,19 @@ namespace ProjectClock.MVC.Controllers
             {
                 return View(dto);
             }
-          
+            if (!HttpContext.User.Claims.TryGetAuthenticatedUserId(out var userId))
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
+            dto.UserId = userId;
+
             var resultDto = await _accountService.EditAccountPassword(dto);
 
             if (resultDto.EditPasswordFailed)
             {
-                return View(resultDto);
+                dto.Result = resultDto;
+                return View(dto);
             }
 
             return RedirectToAction("Index", "Home");
@@ -160,18 +152,8 @@ namespace ProjectClock.MVC.Controllers
 
         [Authorize(Roles = "User")]
         public IActionResult Delete()
-        {
-            if (!HttpContext.User.Claims.TryGetAuthenticatedUserId(out var userId))
-            {
-                return RedirectToAction("Index", "Home");
-            }
-
-            var dto = new DeleteAccountDto()
-            {
-                Id = userId
-            };
-
-            return View(dto);
+        {      
+            return View();
         }
 
         [HttpPost]
@@ -182,6 +164,13 @@ namespace ProjectClock.MVC.Controllers
             {
                 return View(dto);
             }
+            if (!HttpContext.User.Claims.TryGetAuthenticatedUserId(out var userId))
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
+            dto.Id = userId;
+
 
             var deletionSuccessful = await _accountService.DeleteAccount(dto);
